Plugins: categorize linux.check_afinfo as a malware plugin + test fix
