+++ test/plugins/linux/linux.py
@@ -426,7 +426,7 @@ class TestLinuxPageCacheInodepages:
 class TestLinuxCheckAfinfo:
     def test_linux_generic_check_afinfo(self, image, volatility, python):
         rc, out, _err = test_volatility.runvol_plugin(
-            "linux.check_afinfo.Check_afinfo", image, volatility, python
+            "linux.malware.check_afinfo.Check_afinfo", image, volatility, python
         )
 
         # linux-sample-1.bin has no suspicious results.
